+++ include/grid.h
@@ -13,13 +13,13 @@ namespace grid {
         vector<double> xs, ys, zs;
         int nx, ny, nz, nxy, nxyz, nvar;
 
-        float3 At(int3 idx);
+        double3 At(int3 idx);
         int GetIndex(int4 idx);
         int GetIndex(int3 idx, int dir = 0);
         int GetIndex(int i, int j, int k, int d = 0);
         int4 GetIndex(int n);
-        int3 FindIndex(float3 pos, float epsi);
-        vector<int3> ParsePort(float3 src, float3 dst, float epsi);
+        int3 FindIndex(double3 pos, float epsi);
+        vector<int3> ParsePort(double3 src, double3 dst, float epsi);
     };
 }
